Base/App: fix warnings from code analysers: * convert old-style-casts to explicit C++ casts where possible * make some implicit conversions explicit

+++ src/Base/FutureWatcherProgress.cpp
@@ -40,7 +40,7 @@ void FutureWatcherProgress::progressValueChanged(int v)
 {
     if (steps == 0)
         return;
-    unsigned int step = (100 * v) / steps;
+    unsigned int step = (100 * static_cast<unsigned int>(v)) / steps;
     if (step > current) {
         current = step;
         seq.next();
